Discovery: Move discovery classes into their own namespace.
