+++ collabora_online.services.yml
@@ -5,10 +5,10 @@ services:
   logger.channel.collabora_online:
     parent: logger.channel_base
     arguments: ['cool']
-  Drupal\collabora_online\Cool\CollaboraDiscoveryFetcherInterface:
-    class: Drupal\collabora_online\Cool\CollaboraDiscoveryFetcher
-  Drupal\collabora_online\Cool\CollaboraDiscoveryInterface:
-    class: Drupal\collabora_online\Cool\CollaboraDiscovery
+  Drupal\collabora_online\Discovery\CollaboraDiscoveryFetcherInterface:
+    class: Drupal\collabora_online\Discovery\CollaboraDiscoveryFetcher
+  Drupal\collabora_online\Discovery\CollaboraDiscoveryInterface:
+    class: Drupal\collabora_online\Discovery\CollaboraDiscovery
   Drupal\collabora_online\Jwt\JwtTranscoderInterface:
     class: Drupal\collabora_online\Jwt\JwtTranscoder
   Drupal\collabora_online\MediaHelperInterface:
