+++ src/Discovery/CollaboraDiscoveryInterface.php
@@ -12,7 +12,7 @@
 
 declare(strict_types=1);
 
-namespace Drupal\collabora_online\Cool;
+namespace Drupal\collabora_online\Discovery;
 
 /**
  * Service to get a WOPI client URL for a given MIME type.
